+++ FinalSubmission/src/BBox.cpp
@@ -52,13 +52,13 @@ bool BBox::intersect(const Ray &_ray)
   //using the slab method for bbox intersections,
   // an explanation can be found at: https://tavianator.com/fast-branchless-raybounding-box-intersections/
   //first two parametric values are set to theur numerical limits
-  float tMin = std::numeric_limits<float>::min();
-  float tMax = std::numeric_limits<float>::max();
+  double tMin = std::numeric_limits<double>::min();
+  double tMax = std::numeric_limits<double>::max();
   //it then runs through each of x, y and z and the ray is checked for intersections with each pair of faces
   for (int i: {0, 1, 2}){
     //finding the t value for each of the slabs
-    float t1 = (m_pMin[i] - _ray.m_origin[i]) * _ray.m_invDirection[i];
-    float t2 = (m_pMax[i] - _ray.m_origin[i]) * _ray.m_invDirection[i];
+    double t1 = (m_pMin[i] - _ray.m_origin[i]) * _ray.m_invDirection[i];
+    double t2 = (m_pMax[i] - _ray.m_origin[i]) * _ray.m_invDirection[i];
     //if the maximum value is greater than the minimum value then the ray intersects that particular slab
     tMin = std::max(tMin, std::min(t1, t2));
     tMax = std::min(tMax, std::max(t1, t2));
