+++ FinalSubmission/src/BRDF.cpp
@@ -2,7 +2,7 @@
 
 #include "BRDF.hpp"
 
-float BRDF::fresnel(ngl::Vec3 _l, ngl::Vec3 _n)
+double BRDF::fresnel(ngl::Vec3 _l, ngl::Vec3 _n)
 {
   //return m_f0 + (1 - m_f0) * (1 - (_l.dot(_n))) ** 5;
 }
