+++ FinalSubmission/src/Light.cpp
@@ -5,7 +5,7 @@
 
 #include "Light.hpp"
 
-Light::Light(ngl::Vec3 _pos, ngl::Colour _colour, float _intensity):
+Light::Light(ngl::Vec3 _pos, ngl::Colour _colour, double _intensity):
   m_pos(_pos),
   m_colour(_colour),
   m_intensity(_intensity)
