+++ FinalSubmission/src/TextureFile.cpp
@@ -29,10 +29,10 @@ void TextureFile::loadImage()
     for (int x = 0; x < m_imgWidth; x++){
       for (int y = 0; y < m_imgHeight; y++){
         colour = image.pixel(x, y);
-        ngl::Colour c((float)qRed(colour)/255.0,
-                      (float)qGreen(colour)/255.0,
-                      (float)qBlue(colour)/255.0,
-                      (float)qAlpha(colour)/255.0);
+        ngl::Colour c((double)qRed(colour)/255.0,
+                      (double)qGreen(colour)/255.0,
+                      (double)qBlue(colour)/255.0,
+                      (double)qAlpha(colour)/255.0);
         m_pixels[x + m_imgWidth * y] = c;
       }
     }
@@ -43,7 +43,7 @@ void TextureFile::loadImage()
   }
 }
 
-ngl::Colour TextureFile::getPixel(float _u, float _v)
+ngl::Colour TextureFile::getPixel(double _u, double _v)
 {
   int x = _u * m_imgWidth;
   int y = (1-_v) * m_imgHeight;
